examples/singerlyricsanalysis/requirements.txt
jieba
numpy
pillow
musicdl==2.3.6
snownlp
wordcloud
pyecharts